ai_automation/uleth_event_notifier/requirements.txt
feedparser
openai
requests
python-slack-sdk
